restructure the hash-module to a python package to have a better useability and independence
