+++ src/backend/hash-module/README.md
@@ -1,0 +1,1 @@
+pip install -e .
