+++ src/backend/hash-module/pyproject.toml
@@ -1,0 +1,15 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "hash_lib"
+version = "0.1.0"
+description = "A modular hashing and cracking library"
+dependencies = [
+    "requests",
+    "cryptography",
+]
+
+[tool.setuptools.packages.find]
+where = ["src"]
