+++ src/backend/hash-module/src/hash_lib/__init__.py
@@ -1,0 +1,3 @@
+from .hash_core.hasher import Hasher
+from .hash_cracker.cracker import Cracker
+from .fetcher.fetchter import Fetcher
