+++ src/backend/hash-module/src/hash_lib/fetcher/fetchter.py
@@ -1,0 +1,3 @@
+
+class Fetcher:
+    pass
